feat: migrate project to TypeScript and restructure codebase
chore: remove runners folder

chore: remove devmode

refactor: drop typescript and build tools

fix: use 'git' as default bin

feat: add package managers

refactor: simplify runner invocation

chore: restore .gitignore

chore: use `src` in export declaration
