+++ bin/pnpm.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+import('../src/index.js').then(module => module.run('pnpm'));
