+++ readme.md
@@ -19,3 +19,15 @@ yarn global add @danakt/pshe
 ```sh
 npm install -g @danakt/pshe
 ```
+
+## Разработка
+
+### Установка
+
+```sh
+pnpm install
+```
+
+### Переменные окружения
+
+- `PSHE_ISMAC` - Принудительно включить режим раскладки Mac (значения: 'true', '1'). Если не установлено, определяется автоматически на основе платформы.
